Remove receive fragment and move logic to ReceiveActivity
