+++ app/src/main/java/blockeq/com/stellarwallet/activities/ReceiveActivity.kt
@@ -1,0 +1,50 @@
+package blockeq.com.stellarwallet.activities
+
+import android.content.ClipData
+import android.content.ClipboardManager
+import android.content.Context
+import android.os.Bundle
+import android.widget.ImageView
+import android.widget.Toast
+import blockeq.com.stellarwallet.R
+import com.google.zxing.BarcodeFormat
+import com.journeyapps.barcodescanner.BarcodeEncoder
+import kotlinx.android.synthetic.main.content_receive.*
+
+class ReceiveActivity : BasePopupActivity() {
+
+    val address = "GD5HQAPT5KOIKMY35QREYSS34BC3O4FFNTE2DTXUZI4YSJSUXP5QRQS3"
+
+    override fun setTitle(): Int {
+        return R.string.receive_title
+    }
+
+    override fun setContent(): Int {
+        return R.layout.content_receive
+    }
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+        address_text.text = address
+
+        generateQRCode(address, address_qrcode, 500)
+
+        copy_image_button.setOnClickListener { copyDataToClipBoard(address)  }
+    }
+
+    private fun generateQRCode(data: String, imageView: ImageView, size: Int) {
+        val barcodeEncoder = BarcodeEncoder()
+        val bitmap = barcodeEncoder.encodeBitmap(data, BarcodeFormat.QR_CODE, size, size)
+        imageView.setImageBitmap(bitmap)
+    }
+
+    private fun copyDataToClipBoard(data: String) {
+        val clipboard = getSystemService(Context.CLIPBOARD_SERVICE) as ClipboardManager
+        val clip = ClipData.newPlainText("BlockEq Address", data)
+        clipboard.primaryClip = clip;
+
+        Toast.makeText(this, getString(R.string.address_copied_message), Toast.LENGTH_LONG).show()
+
+    }
+}
