+++ app/src/main/res/layout/activity_base_popup.xml
@@ -23,15 +23,19 @@
 
             <RelativeLayout
                 android:layout_width="match_parent"
-                android:layout_height="match_parent">
+                android:layout_height="wrap_content">
 
                 <TextView
                     android:id="@+id/titleText"
-                    style="@style/NavTitleText"
                     android:layout_width="wrap_content"
-                    android:layout_height="match_parent"
-                    android:layout_centerInParent="true"
-                    android:gravity="center" />
+                    android:layout_height="wrap_content"
+                    style="@style/TitleText"
+                    android:layout_marginStart="@dimen/activity_vertical_margin"
+                    android:layout_marginTop="@dimen/activity_vertical_margin"
+                    android:layout_centerVertical="true"
+                    android:textSize="28sp"
+                    android:textColor="@color/black"
+                    android:layout_alignParentStart="true"/>
 
                 <ImageButton
                     android:id="@+id/cancelButton"
@@ -40,7 +44,7 @@
                     android:layout_height="match_parent"
                     android:layout_alignParentEnd="true"
                     android:src="@drawable/ic_close_black_24dp"
-                    android:tint="@color/colorPrimary" />
+                    android:tint="@color/colorAccent" />
             </RelativeLayout>
         </android.support.v7.widget.Toolbar>
