+++ app/src/main/res/layout/fragment_wallet.xml
@@ -25,7 +25,7 @@
                 android:layout_height="wrap_content">
 
                 <Button
-                    android:id="@+id/recieveButton"
+                    android:id="@+id/receiveButton"
                     style="@style/ButtonStyle"
                     android:layout_width="wrap_content"
                     android:layout_height="match_parent"
