+++ app/src/main/res/menu/navigation.xml
@@ -12,11 +12,6 @@
         android:title="@string/nav_trading"/>
 
     <item
-        android:id="@+id/nav_receive"
-        android:icon="@drawable/ic_receive"
-        android:title="@string/nav_receive"/>
-
-    <item
         android:id="@+id/nav_settings"
         android:icon="@drawable/ic_settings"
         android:title="@string/nav_settings"/>
